refactor: Move `load_model_setup` to separate `model_setup` module
