+++ src/coral_models/finetune.py
@@ -10,28 +10,10 @@ from transformers import EarlyStoppingCallback, TrainerCallback, TrainingArgumen
 from transformers.trainer import OptimizerNames
 
 from .data import clean_dataset, load_data
+from .model_setup import load_model_setup
 from .protocols import ModelSetup
-from .wav2vec2 import Wav2Vec2ModelSetup
 
 logger = logging.getLogger(__name__)
-
-
-def load_model_setup(cfg: DictConfig) -> ModelSetup:
-    """Get the model setup for the given configuration.
-
-    Args:
-        cfg (DictConfig):
-            The Hydra configuration object.
-
-    Returns:
-        ModelSetup:
-            The model setup.
-    """
-    model_type: str = cfg.model.type
-    if model_type == "wav2vec2":
-        return Wav2Vec2ModelSetup(cfg)
-    else:
-        raise ValueError(f"Unsupported model type: {model_type!r}")
 
 
 def finetune(cfg: DictConfig) -> None:
